backend/requirements.txt
Django==3.2
djangorestframework==3.12.4
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
djoser==2.2.0
Pillow==10.0.0
progress==1.6
python-dotenv==1.0.0
psycopg2-binary==2.9.3